+++ tests/testthat/test-estimateParams.R
@@ -17,8 +17,7 @@ test_that("prepareData works as expected", {
 
 # negLogLik ----
 
-test_that("negLogLik works as expected", {
-  # Exponential transition
+test_that("negLogLik works as expected for Exponential", {
   transition <- exponential_transition(2, 1.3, 0.8)
   data <- prepareData(getClinicalTrials(
     nRep = 1, nPat = 50, seed = 1234, datType = "1rowPatient",
@@ -27,7 +26,9 @@ test_that("negLogLik works as expected", {
   )[[1]])
   actual1 <- negLogLik(transition, data)
   expect_equal(actual1, 58.65772)
-  # Weibull transition
+})
+
+test_that("negLogLik works as expected for Weibull", {
   transition <- weibull_transition(h01 = 0.2, h02 = 0.5, h12 = 1.6, p01 = 1, p02 = 2.5, p12 = 3)
   data <- prepareData(getClinicalTrials(
     nRep = 1, nPat = 50, seed = 1234, datType = "1rowPatient",
@@ -40,12 +41,13 @@ test_that("negLogLik works as expected", {
 
 # haz ----
 
-test_that("haz works as expected", {
-  # Exponential transition
+test_that("haz works as expected for Exponential", {
   transition <- exponential_transition(h01 = 1.2, h02 = 1.5, h12 = 1.6)
   actual1 <- haz(transition, 0.4, 2)
   expect_equal(actual1, 1.5)
-  # Weibull transition
+})
+
+test_that("haz works as expected for Weibull", {
   transition <- weibull_transition(h01 = 1.2, h02 = 1.5, h12 = 1.6, p01 = 2, p02 = 2.5, p12 = 3)
   actual2 <- haz(transition, 0.4, 2)
   expect_equal(actual2, 0.9486833)
@@ -53,12 +55,13 @@ test_that("haz works as expected", {
 
 # survTrans----
 
-test_that("survTrans works as expected", {
-  # Exponential transition
+test_that("survTrans works as expected for Exponential", {
   transition <- exponential_transition(h01 = 1.2, h02 = 1.5, h12 = 1.6)
   actual1 <- survTrans(transition, 0.4, 2)
   expect_equal(actual1, 0.54881164)
-  # Weibull transition
+})
+
+test_that("survTrans works as expected for Weibull", {
   transition <- weibull_transition(h01 = 1.2, h02 = 1.5, h12 = 1.6, p01 = 2, p02 = 2.5, p12 = 3)
   actual2 <- survTrans(transition, 0.1, 3)
   expect_equal(actual2, 0.99840128)
@@ -66,12 +69,13 @@ test_that("survTrans works as expected", {
 
 # getInit ----
 
-test_that("getInit works as expected", {
-  # Exponential transition
+test_that("getInit works as expected for Exponential", {
   transition <- exponential_transition(h01 = 2.2, h02 = 0.5, h12 = 1.3)
   actual1 <- getInit(transition)
   expect_equal(actual1, c(2.2, 0.5, 1.3))
-  # Weibull transition
+})
+
+test_that("getInit works as expected for Weibull", {
   transition <- weibull_transition(h01 = 0.2, h02 = 0.5, h12 = 1.6, p01 = 1, p02 = 2.5, p12 = 3)
   actual2 <- getInit(transition)
   expect_equal(actual2, c(0.2, 0.5, 1.6, 1, 2.5, 3))
@@ -79,8 +83,7 @@ test_that("getInit works as expected", {
 
 # getTarget ----
 
-test_that("getTarget works as expected", {
-  # Exponential transition
+test_that("getTarget works as expected for Exponential", {
   transition <- exponential_transition(2, 1.3, 0.8)
   data <- prepareData(getClinicalTrials(
     nRep = 1, nPat = 50, seed = 1234, datType = "1rowPatient",
@@ -88,9 +91,12 @@ test_that("getTarget works as expected", {
     accrual = list(param = "intensity", value = 7)
   )[[1]])
   params <- c(1.2, 1.5, 1.6)
-  actual1 <- getTarget(params, data, transition)
+  target <- getTarget(transition)
+  actual1 <- target(params, data)
   expect_equal(actual1, 84.68301)
-  # Weibull transition
+})
+
+test_that("getTarget works as expected for Weibull", {
   transition <- weibull_transition(h01 = 0.2, h02 = 0.5, h12 = 1.6, p01 = 1, p02 = 2.5, p12 = 3)
   data <- prepareData(getClinicalTrials(
     nRep = 1, nPat = 50, seed = 1234, datType = "1rowPatient",
@@ -98,18 +104,20 @@ test_that("getTarget works as expected", {
     accrual = list(param = "intensity", value = 7)
   )[[1]])
   params <- c(1.2, 1.5, 1.6, 2, 1, 2)
-  actual2 <- getTarget(params, data, transition)
+  target <- getTarget(transition)
+  actual2 <- target(params, data)
   expect_equal(actual2, 103.6357444)
 })
 
 # getResults ----
 
-test_that("getResults works as expected", {
-  # Exponential transition
+test_that("getResults works as expected for Exponential", {
   results <- c(1.2, 1.5, 1.6)
   actual1 <- getResults(exponential_transition(), results)
   expect_identical(actual1$hazards, list(h01 = 1.2, h02 = 1.5, h12 = 1.6))
-  # Weibull transition
+})
+
+test_that("getResults works as expected for weibull", {
   results <- c(1.2, 1.5, 1.6, 2, 1, 0.5)
   actual2 <- getResults(weibull_transition(), results)
   expect_identical(actual2$hazards, list(h01 = 1.2, h02 = 1.5, h12 = 1.6))
@@ -118,8 +126,7 @@ test_that("getResults works as expected", {
 
 # estimateParams ----
 
-test_that("estimateParams estimates the true parameters correctly", {
-  # Exponential transition
+test_that("estimateParams estimates the true parameters correctly for Exponential", {
   transition <- exponential_transition(2, 1.3, 0.8)
   data <- getClinicalTrials(
     nRep = 1, nPat = 100000, seed = 123, datType = "1rowPatient",
@@ -130,7 +137,9 @@ test_that("estimateParams estimates the true parameters correctly", {
   expect_equal(actual1$hazards, list(h01 = 2, h02 = 1.3, h12 = 0.8), tolerance = 1e-2)
   expect_identical(actual1$weibull_rates, list(p01 = 1, p02 = 1, p12 = 1))
   expect_identical(class(actual1), c("ExponentialTransition", "TransitionParameters"))
-  # Weibull transition
+})
+
+test_that("estimateParams estimates the true parameters correctly for Weibull", {
   transition <- weibull_transition(h01 = 0.4, h02 = 0.9, h12 = 1.6, p01 = 1, p02 = 0.5, p12 = 1.9)
   data <- getClinicalTrials(
     nRep = 1, nPat = 100000, seed = 123, datType = "1rowPatient",